requirements.txt
dj-database-url==0.4.2
Django==1.10.6
django-multiselectfield==0.1.5
gunicorn==19.6.0
psycopg2==2.6.2
python-http-client==2.2.1
sendgrid==3.6.3
whitenoise==3.2.3
